feat: Move gen-dir-tree.js and sloc.sh scripts to the /scripts directory
The gen-dir-tree.js and sloc.sh scripts were previously located in the server/scripts directory, but have now been moved to the /scripts directory. This change makes it easier to access the scripts and keeps all project-related scripts in a single location. This is a feat type of commit because it adds new functionality or features to the project.